+++ package.json
@@ -17,6 +17,7 @@
     "cors": "^2.8.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-icons": "^5.2.1",
     "react-loader-spinner": "^6.1.6",
     "react-router-dom": "^6.24.1",
     "react-scrollbars-custom": "^4.1.1"
