--- src/components/Blog.jsx
@@ -1,8 +1,0 @@
-
-function Blog() {
-    return (
-        <div className='h-screen w-full bg-orange-600'></div>
-    )
-}
-
-export default Blog
