--- src/components/Categories_Component/CategoryLayout.jsx
@@ -1,11 +1,0 @@
-import { Outlet } from "react-router-dom";
-
-function CategoryLayout() {
-  return (
-    <div>
-      <Outlet />
-    </div>
-  );
-}
-
-export default CategoryLayout;
