+++ src/components/index.js
@@ -1,9 +1,4 @@
-import Home from "./Home";
 import NavBar from "./NavBar";
-import Blog from "./Blog"
-import ContactUs from "./ContactUs";
-import CategoryLayout from "./Categories_Component/CategoryLayout";
-import Categories from "./Categories_Component/Categories";
-import Category1 from "./Categories_Component/Category1"
+import ContactUs from "../pages/ContactUs";
 
-export {Home, NavBar, Blog, ContactUs, CategoryLayout, Categories, Category1};
+export {NavBar, ContactUs};
