+++ src/pages/ContactUs.jsx
@@ -1,0 +1,36 @@
+
+
+const ContactUs = () => {
+    return (
+        <div className=" w-full h-full py-10 bg-orange-100">
+            <div className=" mx-auto w-1/4 ">
+                <div className="bg-white overflow-hidden shadow rounded-lg">
+                    <div className="px-4 py-5 sm:p-6">
+                        <h2 className="text-3xl font-bold text-gray-800 mb-6">Contact Us</h2>
+                        <form>
+                            <div className="mb-4">
+                                <label htmlFor="name" className="block text-sm font-medium text-gray-700">Name</label>
+                                <input type="text" id="name" name="name" className="mt-1 outline-none block w-full border border-gray-300 rounded-md shadow-sm" />
+                            </div>
+                            <div className="mb-4">
+                                <label htmlFor="email" className="block text-sm font-medium text-gray-700">Email</label>
+                                <input type="email" id="email" name="email" className="mt-1 outline-none block w-full border border-gray-300 rounded-md shadow-sm" />
+                            </div>
+                            <div className="mb-4">
+                                <label htmlFor="message" className="block text-sm font-medium text-gray-700">Message</label>
+                                <textarea id="message" name="message" rows="4" className="mt-1 outline-none block w-full border border-gray-300 rounded-md shadow-sm resize-none"></textarea>
+                            </div>
+                            <div className="mt-6">
+                                <button type="submit" className="inline-flex items-center px-6 py-3 border border-transparent text-base font-medium rounded-md text-white bg-orange-600 hover:bg-orange-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-orange-500">
+                                    Send Message
+                                </button>
+                            </div>
+                        </form>
+                    </div>
+                </div>
+            </div>
+        </div>
+    );
+};
+
+export default ContactUs;
